make backend code typescript
