+++ server/db.ts
@@ -1,17 +1,19 @@
-const Database = require('better-sqlite3');
+import Database from "better-sqlite3";
 
 function initializeDatabase() {
-  const db = new Database('./database.sqlite', { verbose: console.log });
-  db.prepare(`
+  const db = new Database("./database.sqlite", { verbose: console.log });
+  db.prepare(
+    `
     CREATE TABLE IF NOT EXISTS organizations (
       id INTEGER PRIMARY KEY AUTOINCREMENT,
       name TEXT NOT NULL,
       created_at DATETIME DEFAULT CURRENT_TIMESTAMP,
       updated_at DATETIME DEFAULT CURRENT_TIMESTAMP
     );
-  `).run();
+  `
+  ).run();
   // TODO: Add your account and deal tables schemas here
   return db;
 }
 
-module.exports = { initializeDatabase };
+export default initializeDatabase;
