--- server/index.js
@@ -1,27 +1,0 @@
-const express = require('express');
-const cors = require('cors');
-const { initializeDatabase } = require('./db');
-const app = express();
-const port = process.env.PORT || 3000;
-
-/**
- * Welcome to the Fullstack Challenge for the Server! 
- * 
- * This is a basic express server. 
- * You can customize and organize it to your needs. 
- * Good luck!
- */
-const db = initializeDatabase();
-
-app.use(cors());
-app.use(express.json());
-app.use(express.urlencoded({ extended: true }));
-
-app.get('/', (req, res) => {
-  const rows = db.prepare('SELECT * FROM organizations').all();
-  res.json({ message: 'Welcome to the server! 🎉', rows });
-});
-
-app.listen(port, () => {
-  console.log(`Server is running on port ${port}`);
-});
